minor warning cleanup

+++ core/OS/OpenSpy.cpp
@@ -216,7 +216,7 @@ namespace OS {
 		 		cursor = v.value._int;
 		 	}
 			arr = reply.values[0].arr_value.values[1].second;
-			for(int i=0;i<arr.arr_value.values.size();i++) {
+			for(size_t i=0;i<arr.arr_value.values.size();i++) {
 				ret = GetGameByRedisKey(arr.arr_value.values[i].second.value._str.c_str(), redis_ctx);
 				break;
 			}
@@ -257,7 +257,7 @@ namespace OS {
 			if (cursor == 0) break;
 			arr = reply.values[0].arr_value.values[1].second;
 
-			for (int i = 0; i<arr.arr_value.values.size(); i++) {
+			for (size_t i = 0; i<arr.arr_value.values.size(); i++) {
 				ret = GetGameByRedisKey(arr.arr_value.values[i].second.value._str.c_str(), redis_ctx);
 				break;
 			}
